fixing servo issues

+++ cloud/dashboard.py
@@ -11,6 +11,7 @@ import sys
 import os
 import logging
 import logging.handlers
+from decimal import * 
 from credentials import AWS_KEY, AWS_SECRET, REGION
 
 
@@ -71,7 +72,7 @@ def get_Data():
             data["Timestamp"] = str(items[0]["Timestamp"])
             data["temperature"] = int(items[0]["Temperature"])
             data["speed"] = int(items[0]["Speed"])
-            data["rpm"] = int(items[0]["RPM"])
+            data["rpm"] = Decimal(items[0]["RPM"])
             data["temperatureavg"] = float(averages["temperature"])
             data["speedaverage"] = float(averages["speed"])
             data["rpmaverage"] = float(averages["rpm"])
